initial commit
1st draft of twitter feed app - this app reads 3 different twitter feeds
and combines the results and creates stats. Results are returned in JSON
via an MVC app.

This was a code test performed for a job application.

The code is functional but it doesn't read from the correct twitter API
call - I failed to get the OAuth authentication working with the twitter
API.